+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Changelog
 0.2
 ---
 
-  * Added Ppx_deriving.{catch,expand_path,path_of_type_decl}.
+  * New plugins: Enum, Bounded.
+  * Added Ppx_deriving.(expand_path,path_of_type_decl,catch,Arg).
+  * Renamed Ppx.deriving.(typ_of_type_decl→core_type_of_type_decl).
   * Change Ppx_deriving.deriver signature to pass ~options and ~path as labeled arguments.
   * Show: include full module path in output.
   * Show: correctly show nested constructors.